+++ require.php
@@ -2,3 +2,4 @@
 
 require_once 'utils.php';
 require_once 'paths.php';
+require_once path_join(FRONTEND_DIR, 'library.php');